__app__/requirements.txt
azure-functions
aiohttp
pendulum